dialog's date display added

+++ app/src/main/java/com/example/tucha/network/vk/NetworkMessage.kt
@@ -8,7 +8,7 @@ data class NetworkMessage(
     @Json(name = "from_id") val fromId: Int = 0,
     @Json(name = "peer_id") val peerId: Int,
     @Json(name = "out") val out: Int = 0,
-    @Json(name = "date") val date: Int = 0,
+    @Json(name = "date") val date: Long = 0,
     @Json(name = "text") val text: String = "",
     @Json(name = "is_hidden") val isHidden: Boolean = false,
     @Json(name = "important") val isImportant: Boolean = false,
